Avoid compilation warning from assigning a float to the return value of atof().
* subversion/mod_dav_svn/repos.c
  (get_entry): Cast return value of atof() to a float.

Suggested by: brane


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@858502 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/mod_dav_svn/repos.c
@@ -1324,7 +1324,7 @@ static const char *get_entry(apr_pool_t *p, accept_rec *result,
         if (parm[0] == 'q'
             && (parm[1] == '\0' || (parm[1] == 's' && parm[2] == '\0')))
           {
-            result->quality = atof(cp);
+            result->quality = (float) atof(cp);
           }
       }
